cube tracing example
new package serves as example for servo_cr - it should trace the
vertices of a cube with 2 arms forever.

ROS_RATE is defined in lib/defines.h and that might not be the best
place for it. At least while testing everything? Maybe it's ok -
it can be overwritten in a single package if needed. The lib should
define a rate that's pretty safe for all robots everywhere.